Optimize S3 layout of datafiles to reduce persistent throttling
The S3 team proposes to change the Object Store File Layout hash scheme
from 32-bit base64 to 24-bit base2. The reduced character range allows
S3 to autoscale more quickly to match the target workload, which will
reduce the amount of time that workloads see persistent throttling. We
believe this change will benefit all workloads, though the magnitude
will vary from workload to workload.

We ran a test with the old and new schemes on a large internal workload.
With the new scheme, we observed that S3 was able to scale and
permanently eliminate throttling within 25 minutes. The old scheme
continued to see throttling for the duration of the 60 minute run, and
saw throttling again when the workload was repeated.

The reduction in bits from 32-bit to 24-bit is to help reduce the
increase in overall key length, with the hash going from a length of 6
characters to 24 characters. 24 bits allows for 2^24 possible prefixes.
Given S3’s documented per-prefix limits of at least 3,500
PUT/COPY/POST/DELETE or 5,500 GET/HEAD requests per second, we get a
maximum scaling potential of about 58 billion requests per second for
PUT, and 92 billion requests per second for GET.